Move standard E2E tests to their own directory
This provides a bit cleaner organization. I like the idea of having
a hierarchial relationship betwen the standard and visual tests, due to
how the visual tests use steps and page object functions from the
standard tests, but I think it is more usable if they are sibling
directories as changed in this commit.